requirements.txt
autopep8==1.4.4
coverage==4.5.3
Django==2.2.3
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
